+++ src/app/constants-old.js
@@ -1,0 +1,16 @@
+/**
+ * Old constant values with deprecation log messages.
+ * Values should be tuples like [oldValue, deprecationMessage].
+ */
+export const Component = {};
+
+export const DataType = {};
+
+export const FileType = {};
+
+export const CoordinationType = {
+  SPATIAL_LAYERS: [
+    'spatialLayers',
+    'The spatialLayers coordination type was split into multiple coordination types in view config schema version 1.0.1.',
+  ],
+};